requirements.txt
streamlit==1.28.0
google-generativeai
pypdf2==3.0.1
python-docx==0.8.11
python-dotenv==1.0.0
plotly==5.15.0